Warning/error fix for r11099 on Mac 10.6
git-svn-id: http://skia.googlecode.com/svn/trunk@11107 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/BlurTest.cpp
@@ -195,9 +195,9 @@ static int step(int x, SkScalar min, SkScalar max) {
 
 // Implement a Gaussian function with 0 mean and std.dev. of 'sigma'.
 static float gaussian(int x, SkScalar sigma) {
-    float k = SK_Scalar1/(sigma * sqrt(2.0f*SK_ScalarPI));
+    float k = SK_Scalar1/(sigma * sqrtf(2.0f*SK_ScalarPI));
     float exponent = -(x * x) / (2 * sigma * sigma);
-    return k * exp(exponent);
+    return k * expf(exponent);
 }
 
 // Perform a brute force convolution of a step function with a Gaussian.
